refactor index.js v4
