+++ src/containers/language/constants.js
@@ -1,0 +1,4 @@
+export const LANGUAGES_LIST = [
+  { lang: 'UA', value: 0 },
+  { lang: 'EN', value: 1 }
+];
