+++ src/context/constants.js
@@ -1,0 +1,3 @@
+export const SNACKBAR_TYPES = {
+  success: 'success'
+};
