+++ src/pages/home/categories-list/categories-list.js
@@ -4,7 +4,8 @@ import Carousel from 'react-multi-carousel';
 import { useTranslation } from 'react-i18next';
 
 import CategoryItem from './category-item';
-import { RESPONSIVE_CATEGORIES, URL_QUERIES_NAME, countPerPage } from '../../../configs/index';
+import { URL_QUERIES_NAME, countPerPage } from '../../../configs/index';
+import { RESPONSIVE_CATEGORIES } from '../constants';
 
 import { CategoriesContext } from '../../../context/categories/categories-context';
 import { useStyles } from './categories-list.style';
