+++ src/pages/images-constructor/constants.js
@@ -1,0 +1,1 @@
+export const DEFAULT_PRICE = ['1400 ', '50 '];
