+++ src/pages/product-details/constants.js
@@ -8,3 +8,23 @@ export const TOAST_SETTINGS = {
   autoClose: 3000,
   hideProgressBar: true
 };
+
+export const RESPONSIVE_PDP = {
+  superLargeDesktop: {
+    breakpoint: { max: 4000, min: 3000 },
+    items: 5
+  },
+  desktop: {
+    breakpoint: { max: 3000, min: 1024 },
+    items: 3
+  },
+  tablet: {
+    breakpoint: { max: 1146, min: 464 },
+    items: 2
+  },
+  mobile: {
+    breakpoint: { max: 810, min: 0 },
+    items: 1
+  }
+};
+export const ERROR = 'error';
