+++ src/utils/constants.js
@@ -1,0 +1,4 @@
+export const SORT_ASC = 'sortAsc';
+export const SORT_DESC = 'sortDesc';
+export const RATE = 'rate';
+export const POPULARITY = 'popularity';
